WIP: 第一阶段拆分
