+++ app/services/platform/__init__.py
@@ -1,0 +1,1 @@
+"""平台服务入口。"""
