+++ app/services/platform/common/__init__.py
@@ -1,0 +1,2 @@
+"""平台服务核心入口。"""
+
