+++ app/services/platform/windows/__init__.py
@@ -1,0 +1,2 @@
+"""Windows 平台服务。"""
+
