+++ app/services/system.py
@@ -1,529 +1,5 @@
-#   AUTO-MAS: A Multi-Script, Multi-Config Management and Automation Software
-#   Copyright © 2024-2025 DLmaster361
-#   Copyright © 2025-2026 AUTO-MAS Team
+"""系统服务兼容入口。"""
 
-#   This file is part of AUTO-MAS.
+from .platform.common.system import System
 
-#   AUTO-MAS is free software: you can redistribute it and/or modify
-#   it under the terms of the GNU Affero General Public License as
-#   published by the Free Software Foundation, either version 3 of
-#   the License, or (at your option) any later version.
-
-#   AUTO-MAS is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty
-#   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
-#   the GNU Affero General Public License for more details.
-
-#   You should have received a copy of the GNU Affero General Public License
-#   along with AUTO-MAS. If not, see <https://www.gnu.org/licenses/>.
-
-#   Contact: DLmaster_361@163.com
-
-
-import sys
-import ctypes
-import asyncio
-import os
-import psutil
-import subprocess
-import tempfile
-import getpass
-from contextlib import suppress
-from dataclasses import dataclass
-from datetime import datetime
-from pathlib import Path
-from typing import Literal, Optional
-
-def __getattr__(name: str) -> object:
-    """延迟导入 Config，避免 app.services 初始化期间触发 app.core 循环导入。"""
-    if name == "Config":
-        from app.core import Config
-        return Config
-    raise AttributeError(name)
-
-from app.utils import ProcessRunner, get_logger
-
-logger = get_logger("系统服务")
-
-
-@dataclass(frozen=True, slots=True)
-class _ProcessPathScan:
-    """按可执行文件路径扫描进程的结果。"""
-
-    pids: list[int]
-    uncertain_pids: list[int]
-    complete: bool
-
-
-class _SystemHandler:
-    ES_CONTINUOUS = 0x80000000
-    ES_SYSTEM_REQUIRED = 0x00000001
-    countdown = 60
-
-    def __init__(self) -> None:
-        self.power_task: Optional[asyncio.Task] = None
-
-    async def set_Sleep(self, if_allow_sleep: bool) -> None:
-        """
-        设置系统休眠
-
-        Parameters
-        ----------
-        if_allow_sleep: bool
-            是否允许系统休眠
-        """
-
-        if if_allow_sleep:
-            # 设置系统电源状态
-            ctypes.windll.kernel32.SetThreadExecutionState(
-                self.ES_CONTINUOUS | self.ES_SYSTEM_REQUIRED
-            )
-        else:
-            # 恢复系统电源状态
-            ctypes.windll.kernel32.SetThreadExecutionState(self.ES_CONTINUOUS)
-
-    async def set_SelfStart(self, if_self_start: bool) -> None:
-        """
-        设置程序开机自启
-
-        Parameters
-        ----------
-        if_self_start: bool
-            程序是否开机自启
-        """
-
-        # 开发环境不管理需要提权的开机自启任务计划
-        if os.getenv("AUTO_MAS_ENV") == "development":
-            return
-
-        if if_self_start:
-
-            # 创建或更新任务计划
-
-            # 获取当前用户和时间
-            current_user = getpass.getuser()
-            current_time = datetime.now().strftime("%Y-%m-%dT%H:%M:%S")
-
-            # XML 模板
-            xml_content = f"""<?xml version="1.0" encoding="UTF-16"?>
-            <Task version="1.2" xmlns="http://schemas.microsoft.com/windows/2004/02/mit/task">
-                <RegistrationInfo>
-                    <Date>{current_time}</Date>
-                    <Author>{current_user}</Author>
-                    <Description>AUTO-MAS自启动服务</Description>
-                    <URI>\\AUTO-MAS_AutoStart</URI>
-                </RegistrationInfo>
-                <Triggers>
-                    <LogonTrigger>
-                        <StartBoundary>{current_time}</StartBoundary>
-                        <Enabled>true</Enabled>
-                    </LogonTrigger>
-                </Triggers>
-                <Principals>
-                    <Principal id="Author">
-                        <LogonType>InteractiveToken</LogonType>
-                        <RunLevel>HighestAvailable</RunLevel>
-                    </Principal>
-                </Principals>
-                <Settings>
-                    <MultipleInstancesPolicy>IgnoreNew</MultipleInstancesPolicy>
-                    <DisallowStartIfOnBatteries>false</DisallowStartIfOnBatteries>
-                    <StopIfGoingOnBatteries>false</StopIfGoingOnBatteries>
-                    <AllowHardTerminate>false</AllowHardTerminate>
-                    <StartWhenAvailable>true</StartWhenAvailable>
-                    <RunOnlyIfNetworkAvailable>false</RunOnlyIfNetworkAvailable>
-                    <IdleSettings>
-                        <StopOnIdleEnd>false</StopOnIdleEnd>
-                        <RestartOnIdle>false</RestartOnIdle>
-                    </IdleSettings>
-                    <AllowStartOnDemand>true</AllowStartOnDemand>
-                    <Enabled>true</Enabled>
-                    <Hidden>false</Hidden>
-                    <RunOnlyIfIdle>false</RunOnlyIfIdle>
-                    <WakeToRun>false</WakeToRun>
-                    <ExecutionTimeLimit>PT0S</ExecutionTimeLimit>
-                    <Priority>7</Priority>
-                </Settings>
-                <Actions Context="Author">
-                    <Exec>
-                        <Command>{Path.cwd() / 'AUTO-MAS.exe'}</Command>
-                        <Arguments>--auto-start</Arguments>
-                    </Exec>
-                </Actions>
-            </Task>"""
-
-            # 创建临时 XML 文件并执行
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".xml", delete=False, encoding="utf-16"
-            ) as f:
-                f.write(xml_content)
-                xml_file = f.name
-
-            try:
-                result = await ProcessRunner.run_process(
-                    "schtasks",
-                    "/create",
-                    "/tn",
-                    "AUTO-MAS_AutoStart",
-                    "/xml",
-                    xml_file,
-                    "/f",
-                )
-
-                if result.returncode == 0:
-                    logger.success(
-                        f"程序自启动任务计划已创建或更新: {Path.cwd() / 'AUTO-MAS.exe'}"
-                    )
-                else:
-                    logger.error(f"程序自启动任务计划创建或更新失败({result.returncode}):")
-                    logger.error(f"  - 标准输出:{result.stdout}")
-                    logger.error(f"  - 错误输出:{result.stderr}")
-
-            except Exception as e:
-                logger.exception(f"程序自启动任务计划创建或更新失败: {e}")
-
-            finally:
-                # 删除临时文件
-                with suppress(Exception):
-                    Path(xml_file).unlink()
-
-        elif not if_self_start and await self.is_startup():
-
-            try:
-
-                result = await ProcessRunner.run_process(
-                    "schtasks", "/delete", "/tn", "AUTO-MAS_AutoStart", "/f"
-                )
-
-                if result.returncode == 0:
-                    logger.success("程序自启动任务计划已删除")
-                else:
-                    logger.error(f"程序自启动任务计划删除失败({result.returncode}):")
-                    logger.error(f"  - 标准输出:{result.stdout}")
-                    logger.error(f"  - 错误输出:{result.stderr}")
-
-            except Exception as e:
-                logger.exception(f"程序自启动任务计划删除失败: {e}")
-
-    async def set_power(
-        self,
-        mode: Literal[
-            "NoAction",
-            "Shutdown",
-            "ShutdownForce",
-            "Reboot",
-            "Hibernate",
-            "Sleep",
-            "KillSelf",
-            "Logoff",
-        ],
-        from_frontend: bool = False,
-    ) -> None:
-        """
-        执行系统电源操作
-
-        :param mode: 电源操作
-        """
-
-        if sys.platform.startswith("win"):
-
-            if mode == "NoAction":
-
-                logger.info("不执行系统电源操作")
-
-            elif mode == "Shutdown":
-
-                await self.kill_emulator_processes()
-                logger.info("执行关机操作")
-                subprocess.run(["shutdown", "/s", "/t", "0"])
-
-            elif mode == "ShutdownForce":
-                logger.info("执行强制关机操作")
-                subprocess.run(["shutdown", "/s", "/t", "0", "/f"])
-
-            elif mode == "Reboot":
-
-                await self.kill_emulator_processes()
-                logger.info("执行重启操作")
-                subprocess.run(["shutdown", "/r", "/t", "0"])
-
-            elif mode == "Hibernate":
-
-                logger.info("执行休眠操作")
-                subprocess.run(["shutdown", "/h"])
-
-            elif mode == "Sleep":
-
-                logger.info("执行睡眠操作")
-                subprocess.run(
-                    ["rundll32.exe", "powrprof.dll,SetSuspendState", "0,1,0"]
-                )
-
-            elif mode == "KillSelf" and Config.server is not None:
-
-                logger.info("执行退出主程序操作")
-                if not from_frontend:
-                    await Config.send_websocket_message(
-                        id="Main", type="Signal", data={"RequestClose": "请求前端关闭"}
-                    )
-                Config.server.should_exit = True
-
-            elif mode == "Logoff":
-
-                await self.kill_emulator_processes()
-                logger.info("执行注销此账户操作")
-                subprocess.run(["shutdown", "/l"])
-
-        elif sys.platform.startswith("linux"):
-
-            if mode == "NoAction":
-
-                logger.info("不执行系统电源操作")
-
-            elif mode == "Shutdown":
-
-                logger.info("执行关机操作")
-                subprocess.run(["shutdown", "-h", "now"])
-
-            elif mode == "Reboot":
-
-                logger.info("执行重启操作")
-                subprocess.run(["shutdown", "-r", "now"])
-
-            elif mode == "Hibernate":
-
-                logger.info("执行休眠操作")
-                subprocess.run(["systemctl", "hibernate"])
-
-            elif mode == "Sleep":
-
-                logger.info("执行睡眠操作")
-                subprocess.run(["systemctl", "suspend"])
-
-            elif mode == "KillSelf" and Config.server is not None:
-
-                logger.info("执行退出主程序操作")
-                if not from_frontend:
-                    await Config.send_websocket_message(
-                        id="Main", type="Signal", data={"RequestClose": "请求前端关闭"}
-                    )
-                Config.server.should_exit = True
-
-            elif mode == "Logoff":
-
-                logger.info("执行注销此账户操作")
-                subprocess.run(["loginctl", "terminate-user", getpass.getuser()])
-
-    async def _power_task(
-        self,
-        power_sign: Literal[
-            "NoAction",
-            "Shutdown",
-            "ShutdownForce",
-            "Reboot",
-            "Hibernate",
-            "Sleep",
-            "KillSelf",
-            "Logoff",
-        ],
-    ) -> None:
-        """电源任务"""
-
-        await asyncio.sleep(self.countdown)
-        await self.set_power(power_sign)
-
-    async def start_power_task(self):
-        """开始电源任务"""
-
-        if self.power_task is None or self.power_task.done():
-            self.power_task = asyncio.create_task(self._power_task(Config.power_sign))
-            logger.info(
-                f"电源任务已启动, {self.countdown}秒后执行: {Config.power_sign}"
-            )
-            Config.power_sign = "NoAction"
-        else:
-            logger.warning("已有电源任务在运行, 请勿重复启动")
-
-    async def cancel_power_task(self):
-        """取消电源任务"""
-
-        if self.power_task is not None and not self.power_task.done():
-            self.power_task.cancel()
-            try:
-                await self.power_task
-            except asyncio.CancelledError:
-                logger.info("电源任务已取消")
-        else:
-            logger.warning("当前无电源任务在运行")
-            raise RuntimeError("当前无电源任务在运行")
-
-    async def kill_emulator_processes(self):
-        """这里暂时仅支持 MuMu 模拟器"""
-
-        logger.info("正在清除模拟器进程")
-
-        keywords = ["Nemu", "nemu", "emulator", "MuMu"]
-        for proc in psutil.process_iter(["pid", "name"]):
-            try:
-                pname = proc.info["name"].lower()
-                if any(keyword.lower() in pname for keyword in keywords):
-                    proc.kill()
-                    logger.info(f"已关闭 MuMu 模拟器进程: {proc.info['name']}")
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                continue
-
-        logger.success("模拟器进程清除完成")
-
-    async def is_startup(self) -> bool:
-        """判断程序是否已经开机自启"""
-
-        try:
-            result = await ProcessRunner.run_process(
-                "schtasks", "/query", "/tn", "AUTO-MAS_AutoStart"
-            )
-            return result.returncode == 0
-        except Exception as e:
-            logger.exception(f"检查任务计划程序失败: {e}")
-            return False
-
-    # async def get_window_info(self) -> list:
-    #     """获取当前前台窗口信息"""
-
-    #     def callback(hwnd, window_info):
-    #         if win32gui.IsWindowVisible(hwnd) and win32gui.GetWindowText(hwnd):
-    #             _, pid = win32process.GetWindowThreadProcessId(hwnd)
-    #             process = psutil.Process(pid)
-    #             window_info.append((win32gui.GetWindowText(hwnd), process.exe()))
-    #         return True
-
-    #     window_info = []
-    #     win32gui.EnumWindows(callback, window_info)
-    #     return window_info
-
-    async def kill_process(
-        self, path: Path | str, *, kill_tree: bool = True
-    ) -> bool:
-        """根据路径中止进程。
-
-        Args:
-            path (Path | str): 目标进程路径。
-            kill_tree (bool): 是否同时中止子进程树。
-
-        Returns:
-            bool: 所有匹配进程均成功中止时返回 True。
-        """
-
-        path = Path(path)
-        logger.info(f"开始中止进程: {path}")
-
-        scan = await self._scan_processes_by_path(path)
-        success = scan.complete
-        first_error: Exception | None = None
-        if scan.uncertain_pids:
-            logger.warning(
-                f"存在无法确认路径的同名进程: {path.name}, PID: {scan.uncertain_pids}"
-            )
-
-        for pid in scan.pids:
-            try:
-                pid_success = await self.kill_process_by_pid(pid, kill_tree=kill_tree)
-            except Exception as e:
-                pid_success = False
-                if first_error is None:
-                    first_error = e
-                logger.opt(exception=True).warning(
-                    f"进程中止异常 PID: {pid}, 原因: {e}"
-                )
-            if not pid_success:
-                success = False
-
-        if first_error is not None:
-            raise first_error
-        if success:
-            logger.success(f"进程已中止: {path}")
-        return success
-
-    async def kill_process_by_pid(self, pid: int, *, kill_tree: bool = True) -> bool:
-        """根据 PID 中止进程。
-
-        Args:
-            pid (int): 目标进程 PID。
-            kill_tree (bool): 是否同时中止子进程树。
-
-        Returns:
-            bool: taskkill 成功执行时返回 True。
-        """
-
-        logger.info(f"开始中止进程 PID: {pid}")
-        args = ["taskkill", "/F"]
-        if kill_tree:
-            args.append("/T")
-        args.extend(["/PID", str(pid)])
-        result = await ProcessRunner.run_process(
-            *args,
-        )
-        if result.returncode != 0:
-            if not psutil.pid_exists(pid):
-                logger.info(f"进程已自行退出 PID: {pid}")
-                return True
-
-            output = result.stderr.strip() or result.stdout.strip() or "无错误信息"
-            logger.warning(
-                f"进程中止失败 PID: {pid}, 返回码: {result.returncode}, 原因: {output}"
-            )
-            return False
-
-        logger.success(f"进程已中止 PID: {pid}")
-        return True
-
-    async def _scan_processes_by_path(self, path: Path | str) -> _ProcessPathScan:
-        """扫描精确路径进程，并记录无法确认路径的同名候选。"""
-
-        path = Path(path)
-        pids: list[int] = []
-        uncertain_pids: list[int] = []
-        complete = True
-        target_path = str(path).casefold()
-        target_name = path.name.casefold()
-
-        try:
-            processes = psutil.process_iter(["pid", "name", "exe"])
-            for proc in processes:
-                try:
-                    info = proc.info
-                except (psutil.NoSuchProcess, psutil.ZombieProcess):
-                    continue
-
-                process_path = info.get("exe")
-                if process_path:
-                    if str(process_path).casefold() == target_path:
-                        pids.append(info["pid"])
-                    continue
-
-                process_name = info.get("name")
-                if process_name and str(process_name).casefold() == target_name:
-                    uncertain_pids.append(info["pid"])
-        except (psutil.AccessDenied, OSError) as e:
-            complete = False
-            logger.warning(f"扫描进程路径失败: {e}")
-
-        return _ProcessPathScan(
-            pids=pids,
-            uncertain_pids=uncertain_pids,
-            complete=complete,
-        )
-
-    async def search_pids(self, path: Path | str) -> list[int]:
-        """
-        根据路径查找进程PID
-
-        :param path: 进程路径
-        :return: 匹配的进程PID列表
-        """
-
-        logger.info(f"开始查找进程 PID: {path}")
-
-        return (await self._scan_processes_by_path(path)).pids
-
-
-System = _SystemHandler()
+__all__ = ["System"]
