+++ app/utils/OCR/OCRtool.py
@@ -1,16 +1,14 @@
 # ocr_tool.py
-import ctypes
 import cv2
 import pyautogui
 from PIL import Image
-import win32con
-import win32gui
 from rapidocr_onnxruntime import RapidOCR
 from mss import mss
 import subprocess
 from pathlib import Path
 
 from app.utils import get_logger
+from app.utils.platform import window as platform_window
 from app.utils.exception import (
     WindowsNotFoundException,
     WindowsNotFocusException,
@@ -100,27 +98,16 @@ class OCRTool:
     def get_system_dpi_scaling(self) -> float:
         """
         获取主显示器的 DPI 缩放比例。
-        兼容 Python 3.13 + 最新 pywin32，使用 ctypes 直接调用 Windows API。
+        由平台窗口能力提供，非 Windows 平台回退到默认缩放比例。
         """
         try:
-            # 启用高 DPI 感知，避免始终返回 96
-            try:
-                ctypes.windll.shcore.SetProcessDpiAwareness(2)  # type: ignore[attr-defined]
-            except (AttributeError, OSError):
-                # Windows 8.1 以下系统没有该函数
-                pass
+            scaling = platform_window.get_dpi_scaling()
+            if scaling is None:
+                raise RuntimeError("系统未返回 DPI 信息")
 
-            # 获取主显示器 DC
-            hdc = win32gui.GetDC(0)
-            try:
-                # 使用 ctypes 直接调用 GetDeviceCaps
-                # LOGPIXELSX = 88
-                dpi = ctypes.windll.gdi32.GetDeviceCaps(hdc, 88)
-            finally:
-                win32gui.ReleaseDC(0, hdc)
-
-            scaling = dpi / 96.0
-            logger.debug(f"检测到系统 DPI: {dpi}, 缩放比例: {scaling:.2f}")
+            logger.debug(
+                f"检测到系统 DPI: {round(scaling * 96)}, 缩放比例: {scaling:.2f}"
+            )
             return scaling
 
         except Exception as e:
@@ -154,17 +141,16 @@ class OCRTool:
         import time
 
         # 查找窗口句柄
-        hwnd = cls._find_window_with_win32gui(title)
+        hwnd = cls._find_window_by_title(title)
         if hwnd is None:
             raise WindowsNotFoundException(f"未能找到标题包含 [{title}] 的窗口")
 
         # 激活窗口并获取区域
         try:
             # 检查窗口状态，只有在最小化时才恢复
-            placement = win32gui.GetWindowPlacement(hwnd)
-            if placement[1] == win32con.SW_SHOWMINIMIZED:
+            if platform_window.is_minimized(hwnd):
                 # 窗口是最小化状态，需要恢复
-                win32gui.ShowWindow(hwnd, win32con.SW_RESTORE)
+                platform_window.restore_window(hwnd)
                 time.sleep(0.05)
 
             # 强制激活窗口
@@ -175,11 +161,11 @@ class OCRTool:
             cls._verify_window_activated(hwnd, title)
 
             # 获取窗口区域
-            rect = win32gui.GetWindowRect(hwnd)
+            rect = platform_window.get_window_rect(hwnd)
             left, top, right, bottom = rect
             region = (left, top, right - left, bottom - top)
 
-            window_title = win32gui.GetWindowText(hwnd)
+            window_title = platform_window.get_window_text(hwnd)
             logger.info(f"成功激活窗口到前台: {window_title} (句柄: {hwnd})")
 
         except WindowsNotFocusException:
@@ -227,7 +213,7 @@ class OCRTool:
         """
         import time
 
-        foreground_hwnd = win32gui.GetForegroundWindow()
+        foreground_hwnd = platform_window.get_foreground_window()
         if foreground_hwnd == hwnd:
             return
 
@@ -236,10 +222,10 @@ class OCRTool:
         cls._force_activate_window(hwnd)
         time.sleep(0.2)
 
-        foreground_hwnd = win32gui.GetForegroundWindow()
+        foreground_hwnd = platform_window.get_foreground_window()
         if foreground_hwnd != hwnd:
-            window_title = win32gui.GetWindowText(hwnd)
-            foreground_title = win32gui.GetWindowText(foreground_hwnd)
+            window_title = platform_window.get_window_text(hwnd)
+            foreground_title = platform_window.get_window_text(foreground_hwnd)
             raise WindowsNotFocusException(
                 f"无法将窗口 [{title}] 激活到前台。\n"
                 f"目标窗口: {window_title} (句柄: {hwnd})\n"
@@ -252,79 +238,21 @@ class OCRTool:
         """
         强制激活窗口到前台（使用线程输入附着技术）
 
+        出错时不抛出异常，因为可能已经部分成功。
+
         Args:
             hwnd: 窗口句柄
         """
-        try:
-            import win32process
-
-            # 获取当前前台窗口
-            foreground_hwnd = win32gui.GetForegroundWindow()
-
-            # 如果已经是前台窗口，直接返回
-            if foreground_hwnd == hwnd:
-                logger.debug("目标窗口已在前台")
-                return
-
-            # 获取前台窗口的线程ID和进程ID
-            if foreground_hwnd:
-                foreground_thread_id, _ = win32process.GetWindowThreadProcessId(foreground_hwnd)
-            else:
-                foreground_thread_id = 0
-
-            # 获取目标窗口的线程ID和进程ID
-            target_thread_id, _ = win32process.GetWindowThreadProcessId(hwnd)
-
-            # 如果线程ID不同，需要附着输入线程
-            attached = False
-            if foreground_thread_id != target_thread_id and foreground_thread_id != 0:
-                try:
-                    # 附着到前台窗口的输入线程
-                    win32process.AttachThreadInput(foreground_thread_id, target_thread_id, True)
-                    attached = True
-                    logger.debug(f"已附着输入线程: {foreground_thread_id} -> {target_thread_id}")
-                except Exception as e:
-                    logger.warning(f"附着输入线程失败: {e}")
-
-            try:
-                # 显示窗口
-                win32gui.ShowWindow(hwnd, win32con.SW_SHOW)
-
-                # 将窗口置于前台
-                win32gui.SetForegroundWindow(hwnd)
-
-                # 设置焦点
-                try:
-                    win32gui.SetFocus(hwnd)
-                except Exception as e:
-                    logger.debug(f"SetFocus 失败（非致命）: {e}")
-
-                # 激活窗口
-                try:
-                    win32gui.BringWindowToTop(hwnd)
-                except Exception as e:
-                    logger.debug(f"BringWindowToTop 失败（非致命）: {e}")
-
-                logger.info(f"已执行窗口激活操作 (句柄: {hwnd})")
-
-            finally:
-                # 分离输入线程（必须在 finally 中执行，确保一定会分离）
-                if attached:
-                    try:
-                        win32process.AttachThreadInput(foreground_thread_id, target_thread_id, False)
-                        logger.debug("已分离输入线程")
-                    except Exception as e:
-                        logger.warning(f"分离输入线程失败: {e}")
-
-        except Exception as e:
-            logger.error(f"强制激活窗口出错: {e}")
-            # 即使出错也不抛出异常，因为可能已经部分成功
+        if platform_window.force_activate_window(hwnd):
+            logger.info(f"已执行窗口激活操作 (句柄: {hwnd})")
+        else:
+            logger.error(f"强制激活窗口出错 (句柄: {hwnd})")
 
 
     @staticmethod
-    def _find_window_with_win32gui(title: str) -> int | None:
+    def _find_window_by_title(title: str) -> int | None:
         """
-        使用 win32gui 查找包含指定标题的窗口。
+        按标题关键字查找窗口。
 
         Args:
             title (str): 窗口标题关键字（模糊匹配）
@@ -332,27 +260,14 @@ class OCRTool:
         Returns:
             int | None: 找到的窗口句柄，未找到返回 None
         """
-        found_windows = []
-
-        def enum_callback(hwnd, results):
-            if win32gui.IsWindowVisible(hwnd):
-                window_title = win32gui.GetWindowText(hwnd)
-                if title.lower() in window_title.lower():
-                    results.append((hwnd, window_title))
-
-        try:
-            win32gui.EnumWindows(enum_callback, found_windows)
-        except Exception as e:
-            logger.error(f"win32gui 枚举窗口失败: {e}")
+        found = platform_window.find_window_by_title(title)
+        if found is None:
             return None
 
-        if found_windows:
-            # 返回第一个匹配的窗口句柄
-            hwnd, window_title = found_windows[0]
-            logger.info(f"win32gui 找到匹配窗口: {window_title} (句柄: {hwnd})")
-            return hwnd
+        hwnd, window_title = found
+        logger.info(f"找到匹配窗口: {window_title} (句柄: {hwnd})")
+        return hwnd
 
-        return None
 
     @classmethod
     def get_screenshot_with_pc(cls, title: str, should_preprocess: bool = True, region: tuple[int, int, int, int] | None = None) -> Image.Image:
