+++ app/utils/ProcessManager.py
@@ -1,603 +1,23 @@
-#   AUTO-MAS: A Multi-Script, Multi-Config Management and Automation Software
-#   Copyright © 2024-2025 DLmaster361
-#   Copyright © 2025-2026 AUTO-MAS Team
-
-#   This file is part of AUTO-MAS.
-
-#   AUTO-MAS is free software: you can redistribute it and/or modify
-#   it under the terms of the GNU Affero General Public License as
-#   published by the Free Software Foundation, either version 3 of
-#   the License, or (at your option) any later version.
-
-#   AUTO-MAS is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty
-#   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
-#   the GNU Affero General Public License for more details.
-
-#   You should have received a copy of the GNU Affero General Public License
-#   along with AUTO-MAS. If not, see <https://www.gnu.org/licenses/>.
-
-#   Contact: DLmaster_361@163.com
-
-
-import os
-import time
-import psutil
-import asyncio
-import win32api
-import win32gui
-import win32con
-import win32process
-
-from contextlib import suppress
-from dataclasses import dataclass
-from datetime import datetime, timedelta
-from pathlib import Path
-
-from .tools import decode_bytes
-from .constants import CREATION_FLAGS
-
-
-@dataclass
-class ProcessInfo:
-    pid: int | None = None
-    name: str | None = None
-    exe: str | None = None
-    cmdline: list[str] | None = None
-
-
-@dataclass
-class ProcessResult:
-    stdout: str
-    stderr: str
-    returncode: int
-
-
-def match_process(proc: psutil.Process, target: ProcessInfo) -> bool:
-    """检查进程是否与目标进程信息匹配"""
-
-    try:
-        if target.pid is not None and proc.pid != target.pid:
-            return False
-        if target.name is not None and proc.name() != target.name:
-            return False
-        if target.exe is not None and Path(proc.exe()) != Path(target.exe):
-            return False
-        if target.cmdline is not None and proc.cmdline() != target.cmdline:
-            return False
-    except (psutil.NoSuchProcess, psutil.AccessDenied):
-        return False
-
-    return True
-
-
-def is_process_running(process_name: str) -> bool:
-    """检查指定进程名是否正在运行且存在可见窗口"""
-
-    for proc in psutil.process_iter(["name"]):
-        with suppress(psutil.NoSuchProcess, psutil.AccessDenied):
-            if proc.info.get("name") == process_name:
-                for hwnd in get_window_handles(proc.pid):
-                    if win32gui.IsWindowVisible(hwnd):
-                        return True
-    return False
-
-
-def get_window_handles(pid: int) -> list[int]:
-    """获取指定进程的所有窗口句柄"""
-
-    window_handles = []
-
-    def enum_callback(hwnd: int, lparam: int) -> bool:
-        """枚举窗口的回调函数"""
-        _, process_id = win32process.GetWindowThreadProcessId(hwnd)
-        if process_id == pid:
-            window_handles.append(hwnd)
-        return True
-
-    win32gui.EnumWindows(enum_callback, 0)
-    return window_handles
-
-
-def get_main_window_handle(
-    pid: int,
-    window_title: str | None = None,
-    window_class_name: str | None = None,
-) -> int | None:
-    """获取指定进程的主窗口句柄
-
-    优先按标题或类名定位, 若未命中则回退到 PID 下最合适的顶层窗口。
-    """
-
-    # 候选过滤: 仅保留可作为主窗口的顶层窗口
-    handles: list[int] = []
-    for hwnd in get_window_handles(pid):
-        try:
-            if not win32gui.IsWindow(hwnd):
-                continue
-            if win32gui.GetParent(hwnd) not in (0, None):
-                continue
-            if win32gui.GetWindow(hwnd, win32con.GW_OWNER):
-                continue
-
-            ex_style = win32gui.GetWindowLong(hwnd, win32con.GWL_EXSTYLE)
-            if ex_style & win32con.WS_EX_TOOLWINDOW:
-                continue
-        except Exception:
-            continue
-
-        handles.append(hwnd)
-
-    # 主流程: 无候选直接失败
-    if not handles:
-        return None
-
-    # 提示匹配: 按标题或类名进一步过滤
-    hinted_handles: list[int] = []
-    if window_title is not None or window_class_name is not None:
-        for hwnd in handles:
-            try:
-                if window_title is not None:
-                    title = win32gui.GetWindowText(hwnd)
-                    if not title or window_title.lower() not in title.lower():
-                        continue
-
-                if window_class_name is not None:
-                    class_name = win32gui.GetClassName(hwnd)
-                    if (
-                        not class_name
-                        or window_class_name.lower() not in class_name.lower()
-                    ):
-                        continue
-            except Exception:
-                continue
-
-            hinted_handles.append(hwnd)
-
-    # 候选排序: 可见优先, 面积次之, 句柄值作为稳定兜底
-    candidates = hinted_handles if hinted_handles else handles
-    best_hwnd: int | None = None
-    best_score: tuple[int, int, int] | None = None
-
-    for hwnd in candidates:
-        try:
-            visible_score = 1 if win32gui.IsWindowVisible(hwnd) else 0
-        except Exception:
-            visible_score = 0
-
-        try:
-            left, top, right, bottom = win32gui.GetWindowRect(hwnd)
-            area_score = max(0, right - left) * max(0, bottom - top)
-        except Exception:
-            area_score = -1
-
-        score = (visible_score, area_score, -hwnd)
-        if best_score is None or score > best_score:
-            best_score = score
-            best_hwnd = hwnd
-
-    return best_hwnd
-
-
-class ProcessManager:
-    """进程监视器类, 用于跟踪主进程及其所有子进程的状态"""
-
-    def __init__(
-        self, window_title: str | None = None, window_class_name: str | None = None
-    ):
-        super().__init__()
-
-        self.process: asyncio.subprocess.Process | None = None
-        self.target_process: psutil.Process | None = None
-        self.window_title = window_title
-        self.window_class_name = window_class_name
-        self._drain_tasks: list[asyncio.Task[None]] = []
-
-    @property
-    def main_pid(self) -> int | None:
-        """主进程的 PID"""
-
-        if self.target_process is not None:
-            return self.target_process.pid
-        if self.process is not None:
-            return self.process.pid
-        return None
-
-    @property
-    def main_process(self) -> psutil.Process | asyncio.subprocess.Process | None:
-        """主进程对象"""
-
-        if self.target_process is not None:
-            return self.target_process
-        if self.process is not None:
-            return self.process
-        return None
-
-    @property
-    def main_hwnd(self) -> int | None:
-        """主进程的主窗口句柄"""
-
-        if self.main_pid is None:
-            return None
-        return get_main_window_handle(
-            self.main_pid, self.window_title, self.window_class_name
-        )
-
-    async def open_process(
-        self,
-        program: Path | str,
-        *args: str,
-        cwd: Path | None = None,
-        target_process: ProcessInfo | None = None,
-        stdin: int = asyncio.subprocess.DEVNULL,
-        stdout: int = asyncio.subprocess.DEVNULL,
-        stderr: int = asyncio.subprocess.DEVNULL,
-        null_stream_to_pipe: bool = False,
-    ) -> None:
-        """
-        启动子进程并跟踪目标进程
-
-        Args:
-            program (Path | str): 可执行文件路径
-            *args (str): 传递给可执行文件的参数
-            cwd (Path | None): 可选的工作目录, 默认为可执行文件所在目录
-            target_process (ProcessInfo | None): 期望目标进程信息, 用于跟踪主进程及其子进程, 默认为 None 表示跟踪直接启动的子进程
-            stdin (int): 标准输入重定向选项, 默认为 asyncio.subprocess.DEVNULL
-            stdout (int): 标准输出重定向选项, 默认为 asyncio.subprocess.DEVNULL
-            stderr (int): 标准错误重定向选项, 默认为 asyncio.subprocess.DEVNULL
-            null_stream_to_pipe (bool): 若为 True, 将设为 DEVNULL 的 stdout/stderr 替换为一条自动销毁输出的标准流管道。
-        """
-
-        if await self.is_running():
-            raise RuntimeError("无法同时管理多个进程")
-
-        if (
-            target_process is not None
-            and target_process.pid is None
-            and target_process.name is None
-            and target_process.cmdline is None
-            and target_process.exe is None
-        ):
-            raise ValueError("目标进程信息不完整")
-
-        await self.clear()
-
-        # 若指定了 null_stream_to_pipe, 将 stdout/stderr 为 DEVNULL 的流替换为管道, 并在后台消费以防止阻塞
-        drain_streams = []
-        if null_stream_to_pipe:
-            if stdout == asyncio.subprocess.DEVNULL:
-                stdout = asyncio.subprocess.PIPE
-                drain_streams.append("stdout")
-            if stderr == asyncio.subprocess.DEVNULL:
-                stderr = asyncio.subprocess.PIPE
-                drain_streams.append("stderr")
-
-        self.process = await asyncio.create_subprocess_exec(
-            program,
-            *args,
-            cwd=cwd or (Path(program).parent if Path(program).is_file() else None),
-            stdin=stdin,
-            stdout=stdout,
-            stderr=stderr,
-            creationflags=CREATION_FLAGS,
-        )
-
-        # 启动协程消费管道流以防止阻塞
-        if drain_streams:
-            for name in drain_streams:
-                stream = getattr(self.process, name)
-                if stream is not None:
-                    self._drain_tasks.append(asyncio.create_task(self._drain(stream)))
-
-        if target_process is not None:
-
-            await self.search_process(
-                target_process,
-                datetime.now() + timedelta(seconds=60),
-                min_create_time=time.time(),
-            )
-
-    async def _drain(self, stream: asyncio.StreamReader) -> None:
-        """
-        消费子进程标准流, 丢弃写入, 防止管道背压阻塞子进程。
-
-        Args:
-            stream (asyncio.StreamReader): 子进程的标准流对象
-        """
-
-        try:
-            while _ := await stream.readline():
-                pass
-        except (ValueError, OSError):
-            # 管道读端在子进程退出后被关闭, 属正常终止; 忽略无效句柄冲突
-            pass
-        except Exception:
-            pass
-
-    async def open_protocol(
-        self, protocol_url: str, target_process: ProcessInfo
-    ) -> None:
-        """
-        使用自定义协议启动子进程, 需要目标进程信息进行跟踪
-
-        Args:
-            protocol_url (str): 自定义协议 URL
-            target_process (ProcessInfo): 期望目标进程信息
-        """
-
-        # 使用 os.startfile 或 subprocess 启动协议
-        try:
-            # 在 Windows 上使用 os.startfile 打开协议
-            if os.name == "nt":
-                os.startfile(protocol_url)
-            else:
-                raise NotImplementedError("仅支持 Windows 平台的自定义协议启动")
-        except Exception as e:
-            raise RuntimeError(f"无法启动协议 {protocol_url}: {e}")
-
-        await self.search_process(
-            target_process,
-            datetime.now() + timedelta(seconds=60),
-            min_create_time=time.time(),
-        )
-
-    async def search_process(
-        self,
-        target_process: ProcessInfo,
-        search_end_time: datetime,
-        min_create_time: float | None = None,
-    ) -> None:
-        """查找目标进程
-
-        Args:
-            target_process: 期望目标进程信息
-            search_end_time: 搜索截止时间
-            min_create_time: 进程创建时间下限（epoch 秒），早于该时刻创建的同名进程视为
-                启动前的残留实例并跳过，避免错误跟踪旧进程（留 2 秒容差吸收时间戳偏差）
-        """
-
-        while datetime.now() < search_end_time:
-            for proc in psutil.process_iter(
-                ["pid", "name", "exe", "cmdline", "create_time"]
-            ):
-                try:
-                    if match_process(proc, target_process):
-                        if (
-                            min_create_time is not None
-                            and proc.create_time() < min_create_time - 2
-                        ):
-                            continue
-                        self.target_process = proc
-                        return
-                except (psutil.NoSuchProcess, psutil.AccessDenied):
-                    continue
-            await asyncio.sleep(0.1)
-        else:
-            raise RuntimeError("未能在限定时间内找到目标进程")
-
-    async def is_running(self) -> bool:
-        """检查当前管理的进程是否仍在运行"""
-
-        if self.target_process is not None:
-            return self.target_process.is_running()
-        if self.process is not None:
-            return self.process.returncode is None
-        return False
-
-    async def kill(self) -> None:
-        """停止监视器并中止所有跟踪的进程"""
-
-        if self.target_process is not None and self.target_process.is_running():
-            with suppress(psutil.NoSuchProcess, psutil.AccessDenied):
-                try:
-                    self.target_process.terminate()
-                    await asyncio.get_running_loop().run_in_executor(
-                        None, self.target_process.wait, 3
-                    )
-                except psutil.TimeoutExpired:
-                    self.target_process.kill()
-                    with suppress(psutil.TimeoutExpired):
-                        await asyncio.get_running_loop().run_in_executor(
-                            None, self.target_process.wait, 3
-                        )
-
-        if self.process is not None and self.process.returncode is None:
-            with suppress(ProcessLookupError):
-                try:
-                    self.process.terminate()
-                    await asyncio.wait_for(self.process.wait(), timeout=3)
-                except asyncio.TimeoutError:
-                    self.process.kill()
-                    with suppress(asyncio.TimeoutError):
-                        await asyncio.wait_for(self.process.wait(), timeout=3)
-
-        await self.clear()
-
-    async def clear(self) -> None:
-        """清空跟踪的进程信息"""
-
-        # 清理残留的排水任务, 避免泄漏
-        if self._drain_tasks:
-            for task in self._drain_tasks:
-                if not task.done():
-                    task.cancel()
-            with suppress(asyncio.CancelledError):
-                await asyncio.gather(*self._drain_tasks, return_exceptions=True)
-            self._drain_tasks = []
-
-        self.process = None
-        self.target_process = None
-
-    async def is_visible(self) -> bool:
-        """检查主进程窗口是否可见
-
-        Returns:
-            bool: 窗口是否可见
-        """
-
-        if self.main_hwnd is None:
-            return False
-
-        try:
-            return bool(win32gui.IsWindowVisible(self.main_hwnd))
-        except Exception:
-            return False
-
-    async def show_window(self) -> bool:
-        """显示主进程窗口
-
-        Returns:
-            bool: 操作是否成功
-        """
-
-        if self.main_hwnd is None:
-            return False
-
-        try:
-            win32gui.ShowWindow(self.main_hwnd, win32con.SW_SHOW)
-            return True
-        except Exception:
-            return False
-
-    async def hide_window(self) -> bool:
-        """隐藏主进程窗口
-
-        Returns:
-            bool: 操作是否成功
-        """
-        if self.main_hwnd is None:
-            return False
-
-        try:
-            win32gui.ShowWindow(self.main_hwnd, win32con.SW_HIDE)
-            return True
-        except Exception:
-            return False
-
-    async def minimize_window(self) -> bool:
-        """最小化主进程窗口
-
-        Returns:
-            bool: 操作是否成功
-        """
-
-        if self.main_hwnd is None:
-            return False
-
-        try:
-            win32gui.ShowWindow(self.main_hwnd, win32con.SW_MINIMIZE)
-            return True
-        except Exception:
-            return False
-
-    async def activate_window(self) -> bool:
-        """激活主进程窗口并将其置于前台
-
-        Returns:
-            bool: 操作是否成功
-        """
-
-        hwnd = self.main_hwnd
-        if hwnd is None:
-            return False
-
-        attached = False
-        current_tid = 0
-        foreground_tid = 0
-
-        try:
-
-            # 若当前线程与前台窗口线程不同, 则附加输入以允许激活窗口
-            foreground_hwnd = win32gui.GetForegroundWindow()
-            if foreground_hwnd:
-                foreground_tid, _ = win32process.GetWindowThreadProcessId(
-                    foreground_hwnd
-                )
-            current_tid = win32api.GetCurrentThreadId()
-            if foreground_tid not in (0, current_tid):
-                win32process.AttachThreadInput(current_tid, foreground_tid, True)
-                attached = True
-
-            # 激活窗口并将其置于前台
-            win32gui.ShowWindow(hwnd, win32con.SW_RESTORE)
-            win32gui.BringWindowToTop(hwnd)
-            try:
-                win32gui.SetForegroundWindow(hwnd)
-            except Exception:
-                # 某些系统策略下 SetForegroundWindow 可能被拒绝, 尝试焦点切换降级路径
-                try:
-                    win32gui.SetWindowPos(
-                        hwnd,
-                        win32con.HWND_TOPMOST,
-                        0,
-                        0,
-                        0,
-                        0,
-                        win32con.SWP_NOMOVE | win32con.SWP_NOSIZE,
-                    )
-                    win32gui.SetWindowPos(
-                        hwnd,
-                        win32con.HWND_NOTOPMOST,
-                        0,
-                        0,
-                        0,
-                        0,
-                        win32con.SWP_NOMOVE | win32con.SWP_NOSIZE,
-                    )
-                    win32gui.SetForegroundWindow(hwnd)
-                except Exception:
-                    return False
-            return True
-        except Exception:
-            return False
-
-        finally:
-            # 取消输入附加
-            if attached:
-                with suppress(Exception):
-                    win32process.AttachThreadInput(current_tid, foreground_tid, False)
-
-
-class ProcessRunner:
-    """用于运行子进程并获取结果的实用程序类"""
-
-    @staticmethod
-    async def run_process(
-        program: Path | str,
-        *args: str,
-        cwd: Path | None = None,
-        timeout: float = 60,
-        if_merge_std: bool = False,
-    ) -> ProcessResult:
-        """运行子进程并获取结果"""
-
-        process = await asyncio.create_subprocess_exec(
-            program,
-            *args,
-            cwd=cwd or (Path(program).parent if Path(program).is_file() else None),
-            stdin=asyncio.subprocess.DEVNULL,
-            stdout=asyncio.subprocess.PIPE,
-            stderr=(
-                asyncio.subprocess.STDOUT if if_merge_std else asyncio.subprocess.PIPE
-            ),
-            creationflags=CREATION_FLAGS,
-        )
-        try:
-            stdout, stderr = await asyncio.wait_for(
-                process.communicate(), timeout=timeout
-            )
-        except asyncio.TimeoutError:
-            with suppress(ProcessLookupError):
-                process.kill()
-            await process.wait()
-            raise
-
-        return ProcessResult(
-            stdout=decode_bytes(stdout),
-            stderr=decode_bytes(stderr),
-            returncode=(
-                process.returncode
-                if process.returncode is not None
-                else await process.wait()
-            ),
-        )
+"""进程能力兼容入口。"""
+
+from .platform.common.process import (
+    ProcessInfo,
+    ProcessManager,
+    ProcessResult,
+    ProcessRunner,
+    get_main_window_handle,
+    get_window_handles,
+    is_process_running,
+    match_process,
+)
+
+__all__ = [
+    "ProcessInfo",
+    "ProcessManager",
+    "ProcessResult",
+    "ProcessRunner",
+    "get_main_window_handle",
+    "get_window_handles",
+    "is_process_running",
+    "match_process",
+]
