+++ app/utils/platform/__init__.py
@@ -1,0 +1,5 @@
+import sys
+
+IS_WINDOWS = sys.platform == "win32"
+
+__all__ = ["IS_WINDOWS"]
