+++ app/utils/platform/common/__init__.py
@@ -1,0 +1,1 @@
+"""通用核心能力。"""
