+++ app/utils/platform/common/errors.py
@@ -1,0 +1,4 @@
+class UnsupportedPlatformError(RuntimeError):
+    def __init__(self, capability: str) -> None:
+        self.capability = capability
+        super().__init__(f"Capability '{capability}' is not supported on this platform")
