+++ app/utils/platform/common/secret.py
@@ -1,0 +1,9 @@
+from app.utils.platform.common.errors import UnsupportedPlatformError
+
+
+def dpapi_encrypt(*args, **kwargs) -> str:
+    raise UnsupportedPlatformError("secret")
+
+
+def dpapi_decrypt(*args, **kwargs) -> str:
+    raise UnsupportedPlatformError("secret")
