+++ app/utils/platform/process.py
@@ -1,0 +1,9 @@
+from . import IS_WINDOWS
+from .common.process_platform import CommonProcessPlatform
+
+if IS_WINDOWS:
+    from .windows.process import WindowsProcessPlatform
+
+    platform_process = WindowsProcessPlatform()
+else:
+    platform_process = CommonProcessPlatform()
