+++ app/utils/platform/secret.py
@@ -1,0 +1,6 @@
+from app.utils.platform import IS_WINDOWS
+
+if IS_WINDOWS:
+    from .windows.secret import *
+else:
+    from .common.secret import *
