+++ app/utils/platform/window.py
@@ -1,0 +1,6 @@
+from . import IS_WINDOWS
+
+if IS_WINDOWS:
+    from .windows.window import *
+else:
+    from .common.window import *
