+++ app/utils/platform/windows/__init__.py
@@ -1,0 +1,2 @@
+"""Windows 平台底层工具。"""
+
