+++ app/utils/platform/windows/secret.py
@@ -1,0 +1,28 @@
+import base64
+import win32crypt
+
+
+def dpapi_encrypt(
+    note: str, description: None | str = None, entropy: None | bytes = None
+) -> str:
+    """使用Windows DPAPI加密数据"""
+
+    if note == "":
+        return ""
+
+    encrypted = win32crypt.CryptProtectData(
+        note.encode("utf-8"), description, entropy, None, None, 0
+    )
+    return base64.b64encode(encrypted).decode("utf-8")
+
+
+def dpapi_decrypt(note: str, entropy: None | bytes = None) -> str:
+    """使用Windows DPAPI解密数据"""
+
+    if note == "":
+        return ""
+
+    decrypted = win32crypt.CryptUnprotectData(
+        base64.b64decode(note), entropy, None, None, 0
+    )
+    return decrypted[1].decode("utf-8")
